- Almost working "test all logs" test

+++ tools/test/TestParseEachFile.py
@@ -2,6 +2,7 @@ import os
 from os import path
 import sys
 import subprocess
+import difflib
 import filecmp
 
 referenceFiles = [
@@ -59,20 +60,29 @@ def compare_two_files(referencefile, testfile):
     # Function compares two files line by line returning status
     status = True
 
-    reference = open(referencefile, "r")
-    test = open(testfile, "r")
-
-    for refline in reference:
-        for testline in test:
-            if not refline == testline:
-                print("Error for line!:", testline, "\n")
+    with open(referencefile) as reference, open(testfile) as testfile:
+        for line1, line2 in zip(reference, testfile):
+            if not line1 == line2:
                 status = False
+
+    # with open(referencefile) as reference:
+    #    reference_text = reference.read()
+    # with open(testfile) as testfile:
+    #    testfile_text = testfile.read()
+    # Find and print the diff:
+    # for line in difflib.unified_diff(reference_text, testfile_text, fromfile=referencefile, tofile=testfile, lineterm=''):
+    #    print(line)
+
+    if not status:
+        print("Mismatch for File:", referencefile, "-->", testfile)
+
     return status
+
 
 
 def prepare_to_test():
     if not len(sys.argv) == 2:
-        print("Error! Provide logger binary file to test!\n")
+        print("Error! Provide logger binary file to test!")
         sys.exit(-1)
     loggerbinaryfile = sys.argv[1]
     # Create default config file
@@ -88,28 +98,25 @@ loggerbinaryfile = sys.argv[1]
 counter = 0
 for x in testFiles:
     testCaseStatus = False
-    print("Test Case: \n", counter)
-    outputFilename = "PARSED" + str(counter) + x
+    print("Test Case:", counter)
+    outputFilename = "PARSED_" + str(counter) + "_" + x
     testFilepath = path.abspath(path.join(currentPath, "..", "..", "example_logs", x))
 
     if not os.path.exists(testFilepath):
-        print("No test file found: \n", testFilepath)
+        print("No test file found:", testFilepath)
         continue
     # Parse file
-    print(loggerbinaryfile, "--parser", "-i", testFilepath, "-o",
-                                outputFilename)
-
-    process = subprocess.Popen([loggerbinaryfile, "--parser", "-i", testFilepath, "-o ",
-                                outputFilename])
+    process = subprocess.Popen([loggerbinaryfile, "--parser", "-i", testFilepath, "-o", outputFilename])
     process.wait()
 
     outputFilepath = os.path.dirname(loggerbinaryfile)
-    fullOutputFilepath = outputFilepath + outputFilename
+    fullOutputFilepath = outputFilename
     referenceFilepath = currentPath + "/reference data/" + referenceFiles[counter]
 
     testCaseStatus = compare_two_files(referenceFilepath, fullOutputFilepath)
 
-    os.remove(fullOutputFilepath)
+    if testCaseStatus:
+        os.remove(fullOutputFilepath)
 
     if testCaseStatus:
         print("Test Case: \n", counter)
